+++ src/mpal/__init__.py
@@ -1,3 +1,3 @@
 """mpal command-line application."""
 
-__version__ = "0.5.2"
+__version__ = "0.6.0"